requirements.txt
Flask==0.11.1
gunicorn==19.6.0
aiohttp==0.22.2
Flask-WTF==0.9.5
Pillow==3.3.0
